# lib/api-zod/src/generated/types/createProductionPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Food Production Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProductionPlanItemsItem } from "./createProductionPlanItemsItem";

export interface CreateProductionPlan {
  planDate: string;
  name: string;
  notes?: string | null;
  items: CreateProductionPlanItemsItem[];
}