# lib/api-zod/src/generated/types/createSubRecipe.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Food Production Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSubRecipeIngredientsItem } from "./createSubRecipeIngredientsItem";

export interface CreateSubRecipe {
  name: string;
  description?: string | null;
  yield: number;
  yieldUnit: string;
  notes?: string | null;
  ingredients: CreateSubRecipeIngredientsItem[];
}