# lib/api-zod/src/generated/types/recipeIngredient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Food Production Planner API
 * OpenAPI spec version: 0.1.0
 */

export interface RecipeIngredient {
  id: number;
  ingredientId: number;
  ingredientName: string;
  unit: string;
  quantity: number;
}