# lib/api-zod/src/generated/types/subRecipeDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Food Production Planner API
 * OpenAPI spec version: 0.1.0
 */
import type { SubRecipe } from "./subRecipe";
import type { SubRecipeIngredient } from "./subRecipeIngredient";

export type SubRecipeDetail = SubRecipe & {
  ingredients: SubRecipeIngredient[];
};